requirements.txt
meshtastic>=2.2.0
paho-mqtt>=1.6.1
PyYAML>=6.0
